Implement CrackerFactory
The idea with this was to get away from the singleton pattern, since it
makes tests difficult and fragile. Dependency injection would have been
an option too, but for simplicity's sake, I chose the factory method.
There were also some strange errors on Windows that I've corrected,
so the program should run on Windows now as well.